prove06

+++ controllers/products-control.js
@@ -5,18 +5,9 @@ const { search } = require("../routes/routes");
 const product = require("../models/product");
 const user = require("../models/user");
 const bcrypt = require("bcryptjs");
+const { validationResult } = require("express-validator");
 const nodemailer = require("nodemailer");
 const sendgridTransport = require("nodemailer-sendgrid-transport");
-
-
-const transporter = nodemailer.createTransport(sendgridTransport({
-    auth: {
-        api_key: "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"
-    }
-}));
-
-let productsSearch = [];
-let previousSearch = "";
 
 const makeProductId = (productName) => {
     return productName + Math.floor((Math.random() * 1000000000)); 
@@ -24,7 +15,7 @@ const makeProductId = (productName) => {
 
 exports.loadSearchPage = (req, res, next) => {
     req.session.previousSearch = req.url;
-    if (productsSearch.length == 0 && previousSearch == "") {
+    if (req.session.foundProducts == 0 && req.session.productSearch == "") {
         Product.find().then(products => {
             res.render("pages/products-search", {
                 title: "Search",
@@ -34,6 +25,7 @@ exports.loadSearchPage = (req, res, next) => {
                 isLoggedIn: req.session.isLoggedIn,
                 isAdmin: (req.user && req.user.isAdmin),
                 errorMessage: req.flash("generalError"),
+                productSearch: req.session.productSearch,
                 csrfT: req.csrfToken()
             });
         }).catch(err => {
@@ -44,18 +36,21 @@ exports.loadSearchPage = (req, res, next) => {
     } else {
         res.render("pages/products-search", {
             title: "Search",
-            products: productsSearch,
+            products: req.session.foundProducts,
             user: req.user,
             cart: false,
             isLoggedIn: req.session.isLoggedIn,
             isAdmin: (req.user && req.user.isAdmin),
             errorMessage: req.flash("generalError"),
+            productSearch: req.session.productSearch,
             csrfT: req.csrfToken()
         });
     }
 }
 
 exports.loadEditPage = (req, res, next) => {
+    let previousForm = req.session.previousForm;
+    req.session.previousForm = {};
     if (req.user && req.user.isAdmin) {
         req.session.previousSearch = req.url;
         if (req.params.productId) {
@@ -68,6 +63,8 @@ exports.loadEditPage = (req, res, next) => {
                         isLoggedIn: req.session.isLoggedIn,
                         isAdmin: (req.user && req.user.isAdmin),
                         errorMessage: req.flash("generalError"),
+                        productSearch: req.session.productSearch,
+                        previousForm: previousForm,
                         csrfT: req.csrfToken()
                     });
                 })
@@ -92,6 +89,8 @@ exports.loadEditPage = (req, res, next) => {
                 isLoggedIn: req.session.isLoggedIn,
                 isAdmin: (req.user && req.user.isAdmin),
                 errorMessage: req.flash("generalError"),
+                productSearch: req.session.productSearch,
+                previousForm: previousForm,
                 csrfT: req.csrfToken()
             });
         }
@@ -106,9 +105,22 @@ exports.loadEditPage = (req, res, next) => {
 }
 
 exports.editProduct = (req, res, next) => {
-    if (req.user && req.user.isAdmin) {
+    let errors = validationResult(req);
+
+    req.session.previousForm = {
+        name: "editProduct",
+        productName: req.body.productName.trim(),
+        productPrice: req.body.productPrice,
+        productDescription: req.body.productDescription.trim(),
+        pictureURL: req.body.pictureURL.trim()
+    }
+
+    if (!errors.isEmpty()) {
+        // There were input validation errors.
+        req.flash("generalError", errors.array()[0].msg);
+        res.redirect(req.session.previousSearch);
+    } else if (req.user && req.user.isAdmin) {
         req.body.productName = req.body.productName.trim();
-        req.body.productPrice = req.body.productPrice.trim();
 
         let newName = (req.body.productName != "" ? req.body.productName : "[-- NO NAME --]");
         let newPrice = (req.body.productPrice > 0 ? req.body.productPrice : 0);
@@ -116,21 +128,29 @@ exports.editProduct = (req, res, next) => {
         if (req.params.productId) {
             // Edit product
             Product.findById(req.params.productId).then(product => {
-                product.name = newName;
-                product.price = newPrice;
-                product.description = req.body.productDescription.trim();
-                product.pictureURL = req.body.pictureURL.trim();
+                if (req.user._id.toString() == product.authorId.toString()) {
+                    product.name = newName;
+                    product.price = parseFloat(newPrice).toFixed(2);
+                    product.description = req.body.productDescription.trim();
+                    product.pictureURL = req.body.pictureURL.trim();
     
-                product.save()
-                .then(result => {
+                    product.save()
+                    .then(result => {
+                        req.session.previousForm = {};
+                        res.redirect("/");
+                    })
+                    .catch(err => {
+                        req.flash("generalError", "Issue saving changes to product.");
+                        res.redirect(req.session.previousSearch);
+                        console.log("Issue saving product. ID: " + req.params.productId + " " + err);
+                    });
+                } else {
+                    req.session.previousForm = {};
+                    req.flash("generalError", "Only the creator of a product can edit it.");
                     res.redirect("/");
-                })
-                .catch(err => {
-                    req.flash("generalError", "Issue saving changes to product.");
-                    res.redirect(req.session.previousSearch);
-                    console.log("Issue saving product. ID: " + req.params.productId + " " + err);
-                });
+                }
             }).catch(err => {
+                req.session.previousForm = {};
                 req.flash("generalError", "Issue getting product.");
                 res.redirect("/");
                 console.log("Issue getting product to update. ID: " + req.params.productId + " " + err);
@@ -140,15 +160,15 @@ exports.editProduct = (req, res, next) => {
             let newProduct = new Product({
                 name: newName,
                 productID: makeProductId(newName),
-                price: newPrice,
+                price: parseFloat(newPrice).toFixed(2),
                 description: req.body.productDescription.trim(),
-                specifications: [],
                 pictureURL: req.body.pictureURL.trim(),
-                tags: []
+                authorId: req.user._id
             });
 
             newProduct.save()
             .then(result => {
+                req.session.previousForm = {};
                 res.redirect("/");
             })
             .catch(err => {
@@ -158,11 +178,13 @@ exports.editProduct = (req, res, next) => {
             });
         }
     } else if (req.session && req.session.isLoggedIn) {
+        req.session.previousForm = {};
         // User not allowed
         req.flash("generalError", "Only admins can edit and add products.");
         res.redirect("/");
     } else {
         // Must sign in to admin account.
+        req.session.previousForm = {};
         res.redirect("/login");
     }
 }
@@ -170,15 +192,28 @@ exports.editProduct = (req, res, next) => {
 exports.removeProduct = (req, res, next) => {
     if (req.user && req.user.isAdmin) {
         console.log("Removing product with _id " + req.params.productId);
-        Product.findByIdAndRemove(req.params.productId)
-            .then(result => {
-                productsSearch = [];
+        Product.findById(req.params.productId)
+        .then(product => {
+            if (req.user._id.toString() == product.authorId.toString()) {
+                Product.findByIdAndRemove(req.params.productId)
+                .then(result => {
+                    req.session.productsSearch = "";
+                    req.session.foundProducts = [];
+                    res.redirect(req.session.previousSearch);
+                })
+                .catch(err => {
+                    req.flash("generalError", "Issue removing product.");
+                    console.log("Failed to remove product with _id " + req.params.productId + ". " + err);
+                });
+            } else {
+                req.flash("generalError", "Only the user who created a product can delete it.");
                 res.redirect(req.session.previousSearch);
-            })
-            .catch(err => {
-                req.flash("generalError", "Issue removing product.");
-                console.log("Failed to remove product with _id " + req.params.productId + ". " + err);
-            });
+            }
+        })
+        .catch(err => {
+            req.flash("generalError", "Issue removing product.");
+            console.log("Failed to remove product with _id " + req.params.productId + ". " + err);
+        });
     } else if (req.session && req.session.isLoggedIn) {
         // User not allowed
         req.flash("generalError", "Only admins can delete products.");
@@ -190,19 +225,24 @@ exports.removeProduct = (req, res, next) => {
 }
 
 exports.searchProducts = (req, res, next) => {
-    productsSearch = [];
+    let errors = validationResult(req);
 
-    Product.find({$or: [
-            {name: {$regex: req.body.searchQuery, $options: 'i'}},
-            {tags: {$regex: req.body.searchQuery, $options: 'i'}}
-        ]
-    }).then(products => {
-        previousSearch = req.body.searchQuery;
-        productsSearch = products;
-        res.redirect("/search");
-    }).catch(err => {
-        req.flash("generalError", "Issue performing search.");
+    if (!errors.isEmpty()) {
+        // There were input validation errors.
+        req.flash("loginError", errors.array()[0].msg);
         res.redirect(req.session.previousSearch);
-        console.log("Issue running product search. " + err );
-    });
+    } else {
+        req.session.productSearch = req.body.searchQuery.trim();
+        req.session.foundProducts = [];
+
+        Product.find({name: {$regex: req.body.searchQuery.trim(), $options: 'i'}
+        }).then(products => {
+            req.session.foundProducts = products;
+            res.redirect("/search");
+        }).catch(err => {
+            req.flash("generalError", "Issue performing search.");
+            res.redirect(req.session.previousSearch);
+            console.log("Issue running product search. " + err );
+        });
+    }
 }
